Introduce `service.type` for all Metricbeat modules (#8965)
The service.type describes which service a dataset is connecting to. This is not necessarily the same for all datasets in one module. An example here is the K8s module which reaches out to 3 different services.

For now the default is just set to the module name if none is set. Unfortunately changing K8s module was not as easy as I though as it still has the old Fetch methods so the variables can not be set.