Honor TEST_ENVIRONMENT in mage integTest (#9663)
If TEST_ENVIRONMENT is set then `mage integTest` will strictly honor its value. If true it will attempt to execute the integration tests and fail with an error if docker is not available. If false it will skip the integTest targets.

Not setting the TEST_ENVIRONMENT value is generally the best option because `mage integTest` will automatically do the right thing. It checks if docker and docker-compose are available and working and then runs the tests. If not available then it will skip the tests (no error, just a notice that they were skipped).