Generate fields.go for netflow input (#9686)
* Generate fields.go for netflow input

Update x-pack/filebeat to generate a fields.go for each input type. And update the code that
generates the include/list.go to import each input.

Also add headers to the modules.d config files.

* Sort imports

* Fix netflow exporter version datatype

It was "int" which is not a valid type. Changed it to long rather than integer because we use that everywhere for numbers.