Add initial Jenkinsfile as foundatin for using Jenkins Pipelines (#9543)
The idea is to move to a Jenkinsfile inside the Beats repository instead of relying in the configs in the JJB files. This should allows us more flexibility in adding / removing jobs directly in PRs to also test new builds. This is inspired / copied from the apm-server Jenkinsfile: https://github.com/elastic/apm-server/blob/master/Jenkinsfile

To make this building a few new minimal JJB files and plugin changes on the Jenkis slave are needed.

This is a very basic Jenkinsfile to get started and only contains the check state. The idea is to first get the setup running and then from there expand the stages we run on Jenkinspipelines. This allows us to test it first without having to run the full test suite always twice. It should also allow a migration step by step by adding a build step to Jenkinsfile and remove it from JJB.