Remove trailing whitespace from golden files (#9827)
When system tests are run with the `GENERATE` environment variable set, we regenerate golden files used by these tests. To do this we use the Python `json.dump` function, which adds trailing whitespaces when it is called with the `indent` argument (which we do). From the Python docs at https://docs.python.org/2/library/json.html:

> Note Since the default item separator is ', ', the output might include trailing whitespace when indent is specified. You can use separators=(',', ': ') to avoid this.

Per the suggestion in the docs, this change adds the `separators` argument as well, to avoid generating trailing whitespace.

This PR also regenerates all golden files with the new regeneration code.