Introduce service.type field for all Filebeat modules (#10042)
Based on ECS the `service.type` field is introduce for all Filebeat modules. By default `service.type` is set to the module name. If the service.type is not the module name the `servic.type` config can be used in a module to overwrite the field.

The `service.name` fields in the Elasticsearch module were removed as `service.name` could be used for example for the cluster name in this context. `service.type` is the field that always stays the same.