[Heartbeat] Fix timing for mode all monitors + Refactor (#10097)
Timing broke in 6.x in #9697 when mode: all is used. Some events will wind up with an empty timestamp, which in go is the year 0001. This isn't slated for release till 6.7 so there's no huge rush here.

This PR seeks to fix not only the bug, but the structural problems in the code that led to the bug. The key issue here is that the timing code should have been centralized for all beats, and invoked in one spot, vs. needing to be invoked in each different dialer type.

That is not a small refactor, but this PR has been able to achieve this and as a result make the code much easier to reason about, and thus safer.

A key thing here was handling the recursive nature of heartbeat tasks in a more uniform way. Previously some functions that modified jobs were recursive, other times they were not. Now, recursive application of modifications only happens in one spot the wrapper.WrapAll and wrapper.Wrap functions (the first one takes a list of jobs, while the second takes a single Job). A new type JobWrapper, can be passed into these. These Job wrapping functions don't handle any recursive logic themselves.

This PR also breaks up some of this utility code, moving it into separate packages. The monitor package became too unwieldy with the additions here.

More tests have been added. There is more work to do to clean stuff up, but there's more than enough here for a single PR.