[Heartbeat] Populate url.port field for http(s) fields. (#10467)
The url.* ECS fields are used for all sorts of stuff. The golang `url` package doesn't infer the URL for URLs with the http(s) schemes, which means that there will be no port value for url.port for these common URLs. This patch explicitly fills in the value in these situations.

This change also fixes some other tests by using an explicit port for TCP which is required in heartbeat. Additionally, it standardizes on the `uint16` type for port values which is common in the go ecosystem.