Cleanup _exists_ queries in dashboards (#10483)
The dashboards for 6.x contained quite a few _exists_ queries to check if some fields. These fields partially do not exists anymore because they were mapped to ECS. Also sometimes _exists_ is used to check some objects. Because of this an automatic renaming based on ecs-migration.yml was not possible. Here is the manual migration. I did tests this with apache module.