Add group_measurements_by_instance_label flag to perfmon configuration (#11002)
This flag will send all perfmon measurements with a matching instance label as part of the same event (i.e. all metrics for C:, Processor X, etc.). This addresses some of the issues raised in #6584.

In most cases enabling this flag considerably reduces the number of events sent by metricbeat.

Co-Authored-By: Josh Smith <j_smith95@live.com>