Export function template (#11923)
A new subcommand is added under `export` named `function`. This command prints the loadable template to stdout for enabled functions.

```
$ ./functionbeat export function cloudwatch
```

I extracted generating templates from `CLIManager`. Now it has a `TemplateBuilder` which constructs the template for the manager. A template builder can be retrieved from a provider, as generating templates is specific to providers. Thus, `CLIManager` does not need to know about the provider anymore.