Add secondary source of modules to Metricbeat to load light modules (#12465)
Implements #12270.

Modify metricbeat modules registry to include an optional secondary modules source.
A modules registry uses the secondary source as a fallback when it cannot provide
the configuration for an specific metricset. Semantics of current methods are
maintained.

An implementation for a secondary source is provided. This implementation loads
modules from disk that depend on traditional registered modules. Once loaded,
a "registration" object can be obtained, what includes a metricset factory built
with the specific overrides needed to instantiate the final metricset.

Beats based on metricbeat can opt-in for this feature by adding
WithLightModules() to its beat.Creator.
This option is added to the licensed metricbeat, for that a new root command
is instantiated.