Look for keystore under the correct path (#13332)
Previously, Functionbeat put the keystore file under `/tmp` due to the overrides of the function. But when loading the keystore file before packaging the Beat, data path was not yet initialized, so it looked for the file under the incorrect folder. Furthermore, the file in the compressed package was hardcoded to `data/functionbeat.keystore`. So there was no way for functionbeat to find the keystore when loaded to the cloud provider.

Closes #13079