Tag image names in metricbeat docker compose files (#13775)
Tag image names so they can be more easily reused, cached and shared.

Modify all Dockerfiles related to metricbeat resources to parameterize
their versions.

Build args are not used anymore in docker compose, and environment is
used instead. The reason is that build args couldn't be used to tag
the images. Build args are still passed to Dockerfiles, and they don't
have default values now, they must be set in the docker compose file.

Some services that are being tested with multiple versions have been
refactorized to be parameterized so they can share most of their
Dockerfiles and docker compose definitions.

Testing framework doesn't force rebuild of images now. If an image
exists with the same tag this image will be used. Before trying to start
a service, it tries to pull its image to avoid building it if it is available.
Both pull and build will reuse local layers if they can.

Part of #13471 and #12909.