Add compose option to advertise hosts for specific ports (#14609)
Docker compose framework for integration tests allows to pass the host
and port exposed in the host network for services that advertise the
address that clients must use to connect to them. This way metricbeat
tests running in the host can connect to them after the initial
negotiation.

Add an option so for services that expose multiple ports, the test can
select what is the port it needs to connect.