Move crawler to filebeat/beater (#16700)
The Crawler is more or less the main run loop of filebeat. It used to
generate the factories for inputs and modules, yet did expose those via
public interfaces, so we can hook them up with autodiscovery.

The crawler constuctor used to create only a half-instantiated object,
requiring all additional setup to be run after crawler.Start, as start
will finally initialized the shared variables.

This change moves the crawler into filebeat and makes it private (it is
not used anywhere in the Beats codebase).
In addition to this, we now create the factories upfront and pass them
as proper dependencies to autodiscovery and the crawler.